Squelette TP2

+++ user/user.h
@@ -25,7 +25,7 @@ int mkdir(const char*);
 int chdir(const char*);
 int dup(int);
 int getpid(void);
-char* sbrk(int);
+char* sbrk(uint64);
 int sleep(int);
 int uptime(void);
 int ntas();
